Panel specific classes
Made each panel it's separate class so ServerWindow woun't have 800 lines by the time it's done.
