+++ sushi/src/main/java/comp1206/sushi/server/DishesPanel.java
@@ -1,0 +1,12 @@
+package comp1206.sushi.server;
+
+import javax.swing.JButton;
+import javax.swing.JPanel;
+
+class DishesPanel extends JPanel {
+
+	  public DishesPanel() {
+	    JButton b1 = new JButton("MCCIES");
+	    add(b1);
+	  }
+	}
