+++ sushi/src/main/java/comp1206/sushi/server/DronesPanel.java
@@ -1,0 +1,12 @@
+package comp1206.sushi.server;
+
+import javax.swing.JButton;
+import javax.swing.JPanel;
+
+class DronesPanel  extends JPanel {
+
+	  public DronesPanel () {
+	    JButton b1 = new JButton("Insert drone stuff");
+	    add(b1);
+	  }
+	}
