+++ sushi/src/main/java/comp1206/sushi/server/IngredientsPanel.java
@@ -1,0 +1,12 @@
+package comp1206.sushi.server;
+
+import javax.swing.JButton;
+import javax.swing.JPanel;
+
+class IngredientsPanel extends JPanel {
+
+	  public IngredientsPanel() {
+	    JButton b1 = new JButton("Add salt");
+	    add(b1);
+	  }
+	}
