+++ sushi/src/main/java/comp1206/sushi/server/OrdersPanel.java
@@ -1,0 +1,12 @@
+package comp1206.sushi.server;
+
+import javax.swing.JButton;
+import javax.swing.JPanel;
+
+class OrdersPanel extends JPanel {
+
+	  public OrdersPanel() {
+	    JButton b1 = new JButton("It will be 1 hr late");
+	    add(b1);
+	  }
+	}
