+++ sushi/src/main/java/comp1206/sushi/server/PostcodesPanel.java
@@ -1,0 +1,12 @@
+package comp1206.sushi.server;
+
+import javax.swing.JButton;
+import javax.swing.JPanel;
+
+class PostcodesPanel extends JPanel {
+
+	  public PostcodesPanel() {
+	    JButton b1 = new JButton("is this working?");
+	    add(b1);
+	  }
+	}
