+++ sushi/src/main/java/comp1206/sushi/server/StaffPanel.java
@@ -1,0 +1,12 @@
+package comp1206.sushi.server;
+
+import javax.swing.JButton;
+import javax.swing.JPanel;
+
+class StaffPanel extends JPanel {
+
+	  public StaffPanel() {
+	    JButton b1 = new JButton("Underpaid students");
+	    add(b1);
+	  }
+	}
