+++ sushi/src/main/java/comp1206/sushi/server/SuppliersPanel.java
@@ -1,0 +1,12 @@
+package comp1206.sushi.server;
+
+import javax.swing.JButton;
+import javax.swing.JPanel;
+
+class SuppliersPanel extends JPanel {
+
+	  public SuppliersPanel() {
+	    JButton b1 = new JButton("Supplier Madness");
+	    add(b1);
+	  }
+	}
