+++ sushi/src/main/java/comp1206/sushi/server/UserPanel.java
@@ -1,0 +1,13 @@
+package comp1206.sushi.server;
+
+import javax.swing.JButton;
+import javax.swing.JPanel;
+
+class UserPanel extends JPanel {
+
+	  public UserPanel() {
+	    JButton b1 = new JButton("You mean students AMRIGHT?");
+	    add(b1);
+	  }
+	}
+
